BASHI-3: Split OpeningTimes view model files | Move Home Page to Blazor
