+++ src/Mandarin/ViewModels/IOpeningTimesViewModel.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Mandarin.ViewModels
+{
+    public interface IOpeningTimesViewModel
+    {
+        IReadOnlyList<OpeningTime> OpeningTimes { get; }
+    }
+}
